requirements.txt
django==3.1.5
Celery==5.0.5
django-crispy-forms==1.10.0
django-qsessions==1.1.2
dj-database-url==0.5.0
django-menu-generator==1.1.0
psycopg2==2.8.6 --no-binary=psycopg2
djangorestframework==3.12.2
django-sql-utils==0.6.1
django-redis==4.12.1
Pillow==8.1.0
django-lifecycle==0.8.0
django-tree-queries==0.4.3
PyJWT==2.0.0
django-sql-utils==0.6.1
